Add optimization for inverse shortest path
The basic optimization is a linear program (see Program 6 in the
write-up). The path set complement is generated by naively
stringing together feasible pairs on paths in the shortest path set.
This works ok in some cases but not others. The failing test case
generated by hypothesis is one such example.

The overall issue with the optimization is that it in general
requires a column generating scheme to guarantee correctness.
In the worst case, this requires enumeration of all of the paths.
This counter example demonstrates undesireble behavior even in a
simple case.