Use entropy in objective function; Refactor
Previous experiments using only the sum of path costs in the
objective resulted in infinite set of global optima. This was a
direct consequence of normalizing the total cost over the out edges
of each node (so that the edge costs relate to the transition
probabilities out of each node). To mitigate this, jointly minimize
the negative total out edge entropy over all nodes and the path cost.

The constraints are the normalizing constraints on the out edges from
each node, the constraint that all paths between the same OD pair are
equal, and tha the cost of each discovered paths is at least that of
the observed path(s) between the corresponding OD pair.
paths are discovered by chaining compatible observed paths together.